[ch23306] [Refactor]interventions redux actions/reducers
