+++ intervention-results/results-structure/modals/cp-output-dialog.ts
@@ -4,7 +4,7 @@ import {sendRequest} from '@unicef-polymer/etools-ajax/etools-ajax-request';
 import {getEndpoint} from '../../../utils/endpoint-helper';
 import {interventionEndpoints} from '../../../utils/intervention-endpoints';
 import {getStore} from '../../../utils/redux-store-access';
-import {getIntervention} from '../../../common/interventions';
+import {getIntervention} from '../../../common/actions/interventions';
 import {fireEvent} from '../../../utils/fire-custom-event';
 import '@unicef-polymer/etools-dropdown/etools-dropdown';
 import '@unicef-polymer/etools-dropdown/etools-dropdown-multi';
