+++ intervention-results/results-structure/modals/indicator-dialog/mixins/save-indicator-mixin.ts
@@ -11,7 +11,7 @@ import {NonClusterIndicatorEl} from '../non-cluster-indicator';
 import {ClusterIndicatorEl} from '../cluster-indicator';
 import {IndicatorDisaggregations} from '../indicator-dissaggregations';
 import {getStore} from '../../../../../utils/redux-store-access';
-import {updateCurrentIntervention} from '../../../../../common/interventions';
+import {updateCurrentIntervention} from '../../../../../common/actions/interventions';
 import {Constructor} from '@unicef-polymer/etools-types';
 
 /**
