+++ intervention-results/results-structure/results-structure.ts
@@ -30,7 +30,7 @@ import {interventionEndpoints} from '../../utils/intervention-endpoints';
 import {pageIsNotCurrentlyActive} from '../../utils/common-methods';
 import '../../common/layout/are-you-sure';
 import get from 'lodash-es/get';
-import {getIntervention, updateCurrentIntervention} from '../../common/interventions';
+import {getIntervention, updateCurrentIntervention} from '../../common/actions/interventions';
 import {_sendRequest} from '../../utils/request-helper';
 import {isUnicefUser, currentIntervention} from '../../common/selectors';
 import cloneDeep from 'lodash-es/cloneDeep';
