+++ intervention-results/supply-agreement/supply-agreement-dialog.ts
@@ -10,7 +10,7 @@ import {getEndpoint} from '../../utils/endpoint-helper';
 import {interventionEndpoints} from '../../utils/intervention-endpoints';
 import {fireEvent} from '../../utils/fire-custom-event';
 import {formatServerErrorAsText} from '@unicef-polymer/etools-ajax/ajax-error-parser';
-import {updateCurrentIntervention} from '../../common/interventions';
+import {updateCurrentIntervention} from '../../common/actions/interventions';
 import '@unicef-polymer/etools-dialog/etools-dialog';
 import '@unicef-polymer/etools-dropdown/etools-dropdown';
 import '@polymer/paper-input/paper-input';
